requirements.txt
fastapi
uvicorn[standard]
langchain
chromadb
python-multipart
pytesseract
pdfplumber
python-docx
beautifulsoup4
requests
langchain-community
langchain-cohere
langchain-groq
cohere
python-dotenv
transformers
